refactor mobile-client signing module: remove legacy signing.ts file and restructure signing logic into modular components, including buildSignedMobileRequest, canonicalization, nonce generation, and payload hashing for improved maintainability and clarity
