+++ packages/mobile-client/src/signing/types.ts
@@ -1,0 +1,66 @@
+import { WalletClientWithAccount } from '@nadohq/shared';
+import { Hex } from 'viem';
+import { MobileSignedRequestParams } from '../types/clientTypes';
+import {
+  MobileNotificationPlatform,
+  MobileServerNotificationPreferences,
+} from '../types/serverTypes';
+
+/**
+ * Unsigned inner payloads that can be authenticated against the mobile service API. `type` must come first
+ * in each literal so msgpack serializes it deterministically (see {@link canonicalizeMobileInner}).
+ */
+export type MobileSignedInner =
+  | { type: 'claim_username'; display_name: string }
+  | { type: 'update_username'; display_name: string }
+  | { type: 'set_private_mode'; private_mode: boolean }
+  | { type: 'self_identity' }
+  | {
+      type: 'register_expo_token';
+      expo_token: string;
+      platform: MobileNotificationPlatform;
+      // Optional on the wire, but always present (null when unset) so the msgpack payload hash matches the
+      // backend, which serializes `Option::None` as nil under the same key.
+      locale: string | null;
+      app_version: string | null;
+    }
+  | { type: 'unregister_expo_token'; expo_token: string }
+  | {
+      type: 'update_preferences';
+      preferences: MobileServerNotificationPreferences;
+    }
+  | { type: 'notification_preferences' }
+  | { type: 'registered_devices' };
+
+/**
+ * Type-specific payload fields for a signed inner request of a given `type` (the inner payload without its
+ * `type` tag). Derived from {@link MobileSignedInner} so the two never drift apart.
+ */
+export type MobileSignedInnerParams<T extends MobileSignedInner['type']> = Omit<
+  Extract<MobileSignedInner, { type: T }>,
+  'type'
+>;
+
+/**
+ * A signed mobile service API request body: the inner payload flattened with its signature, sender, and
+ * nonce (unlike the engine client, which wraps params under a `{ [type]: params }` key).
+ */
+export type MobileSignedRequest<
+  T extends MobileSignedInner = MobileSignedInner,
+> = T & {
+  signature: Hex;
+  sender: Hex;
+  nonce: string;
+};
+
+/**
+ * Params for {@link buildSignedMobileRequest}.
+ */
+export interface BuildSignedMobileRequestParams extends MobileSignedRequestParams {
+  walletClient: WalletClientWithAccount;
+  inner: MobileSignedInner;
+  /**
+   * Overrides the generated nonce; mainly useful for tests that pin the signed payload.
+   */
+  nonce?: bigint;
+}
